Added new services for csv, search, sort and transformation and migrated code into the services
