+++ ui/src/pages/Dashboard/NodeHistory/HistoryInfo.vue
@@ -168,6 +168,10 @@ import axios from 'axios';
 import { MemoryStorage } from 'ttl-localstorage';
 import { ExportToCsv } from 'export-to-csv';
 import VueElementLoading from 'vue-element-loading';
+import CsvService from '../Service/CsvService';
+import SearchService from '../Service/SearchService';
+import TransformationService from '../Service/TransformationService';
+import SortService from '../Service/SortService';
 import {
   httpRequestFluxInfo, httpRequestDaemonInfo, httpRequestFluxHistoryStats,
 } from '../Request/HttpRequest';
@@ -281,25 +285,9 @@ export default {
     processData(sortProps, isProcessingState) {
       let result;
       if (this.searchQuery !== '') {
-        const temp = [];
-        result = this.fuseSearch.search(`=${this.searchQuery}`);
-        for (let i = 0; i < Object.keys(result).length; i += 1) {
-          temp.push(result[i].item);
-        }
-        result = temp;
+        result = SearchService.search(this.fuseSearch, this.searchQuery);
       } else if (this.filters.default.length) {
-        const arr = [];
-        const data = [];
-        this.filters.default.forEach((item) => {
-          const objs = this.filter.get(item);
-          objs.forEach((obj) => {
-            if (!arr.includes(`${obj.roundTime}${obj.total}`)) {
-              arr.push(`${obj.roundTime}${obj.total}`);
-              data.push(obj);
-            }
-          });
-        });
-        result = data;
+        result = TransformationService.processFilters(this.filters, this.filter, 'historyinfo');
       } else {
         result = this.tableData;
       }
@@ -372,136 +360,15 @@ export default {
     },
     setSearch() {
       this.originalData = JSON.stringify(this.tableData);
-      this.fuseSearch = new Fuse(this.tableData, { useExtendedSearch: true, keys: ['roundTime'] });
+      this.fuseSearch = SearchService.generateSearch(Fuse, this.tableData, ['roundTime']);
     },
     sortChange(sortProps) {
       this.processData(sortProps, true);
     },
     sorting(sortProps, data) {
-      if (sortProps.column.label === 'Round Time' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.roundTime > b.roundTime) {
-            val = 1;
-          } else if (a.roundTime < b.roundTime) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Round Time' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.roundTime < b.roundTime) {
-            val = 1;
-          } else if (a.roundTime > b.roundTime) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Round Time Converted' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.roundTime > b.roundTime) {
-            val = 1;
-          } else if (a.roundTime < b.roundTime) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Round Time Converted' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.roundTime < b.roundTime) {
-            val = 1;
-          } else if (a.roundTime > b.roundTime) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Cumulus' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.cumulus > b.cumulus) {
-            val = 1;
-          } else if (a.cumulus < b.cumulus) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Cumulus' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.cumulus < b.cumulus) {
-            val = 1;
-          } else if (a.cumulus > b.cumulus) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Nimbus' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.nimbus > b.nimbus) {
-            val = 1;
-          } else if (a.nimbus < b.nimbus) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Nimbus' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.nimbus < b.nimbus) {
-            val = 1;
-          } else if (a.nimbus > b.nimbus) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Stratus' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.stratus > b.stratus) {
-            val = 1;
-          } else if (a.stratus < b.stratus) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Stratus' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.stratus < b.stratus) {
-            val = 1;
-          } else if (a.stratus > b.stratus) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Total Nodes' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.total > b.total) {
-            val = 1;
-          } else if (a.total < b.total) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Total Nodes' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.total < b.total) {
-            val = 1;
-          } else if (a.total > b.total) {
-            val = -1;
-          }
-          return val;
-        });
-      } else {
-        this.tableData = JSON.parse(this.originalData);
-      }
-      return data;
+      const ret = SortService.sortHistoryInfo(data, sortProps, this.originalData);
+      this.tableData = Object.keys(ret.tableDatas).length > 0 ? ret.tableDatas : this.tableData;
+      return ret.datas;
     },
     processDataForCsv(data) {
       const values = [];
@@ -518,31 +385,16 @@ export default {
       return values;
     },
     downloadCsvFile(data) {
-      const date = new Date();
-      const month = date.getMonth();
-      const day = date.getDate();
-      const year = date.getFullYear();
-      const options = {
-        filename: `History_Info_${month}${day}${year}`,
-        fieldSeparator: ',',
-        quoteStrings: '"',
-        decimalSeparator: '.',
-        showLabels: true,
-        showTitle: true,
-        title: `History_Info - ${month}/${day}/${year}`,
-        useTextFile: false,
-        useBom: true,
-        headers: [
-          'Round Time',
-          'Round Time Converted',
-          'Total Nodes',
-          'Cumulus',
-          'Nimbus',
-          'Stratus',
-        ],
-      };
-      const csvExporter = new ExportToCsv(options);
-      csvExporter.generateCsv(this.processDataForCsv(data));
+      const module = 'History_Info';
+      const headers = [
+        'Round Time',
+        'Round Time Converted',
+        'Total Nodes',
+        'Cumulus',
+        'Nimbus',
+        'Stratus',
+      ];
+      CsvService.Download(this.processDataForCsv(data), headers, module, ExportToCsv);
     },
     processFilters(key) {
       if (!this.filters.default.includes(key)) {
@@ -556,14 +408,7 @@ export default {
       return this.processData(false, false);
     },
     processState(keys) {
-      this.filters.states.map((item) => {
-        const values = item;
-        values.state = false;
-        if (keys.includes(values.name)) {
-          values.state = true;
-        }
-        return values;
-      });
+      this.filters = TransformationService.processState(keys, this.filters);
     },
   },
 };
