+++ ui/src/pages/Dashboard/NodeInfo/NodeApp.vue
@@ -178,6 +178,10 @@ import axios from 'axios';
 import { MemoryStorage } from 'ttl-localstorage';
 import { ExportToCsv } from 'export-to-csv';
 import VueElementLoading from 'vue-element-loading';
+import CsvService from '../Service/CsvService';
+import SearchService from '../Service/SearchService';
+import TransformationService from '../Service/TransformationService';
+import SortService from '../Service/SortService';
 import {
   httpRequestFluxInfo, httpRequestDaemonInfo, httpRequestFluxHistoryStats,
 } from '../Request/HttpRequest';
@@ -296,25 +300,9 @@ export default {
     processData(sortProps, isProcessingState) {
       let result;
       if (this.searchQuery !== '') {
-        const temp = [];
-        result = this.fuseSearch.search(`=${this.searchQuery}`);
-        for (let i = 0; i < Object.keys(result).length; i += 1) {
-          temp.push(result[i].item);
-        }
-        result = temp;
+        result = SearchService.search(this.fuseSearch, this.searchQuery);
       } else if (this.filters.default.length) {
-        const arr = [];
-        const data = [];
-        this.filters.default.forEach((item) => {
-          const objs = this.filter.get(item);
-          objs.forEach((obj) => {
-            if (!arr.includes(`${obj.ip}${obj.apps.count}`)) {
-              arr.push(`${obj.ip}${obj.apps.count}`);
-              data.push(obj);
-            }
-          });
-        });
-        result = data;
+        result = TransformationService.processFilters(this.filters, this.filter, 'nodeapp');
       } else {
         result = this.tableData;
       }
@@ -368,156 +356,15 @@ export default {
     },
     setSearch() {
       this.originalData = JSON.stringify(this.tableData);
-      this.fuseSearch = new Fuse(this.tableData, { useExtendedSearch: true, keys: ['ip'] });
+      this.fuseSearch = SearchService.generateSearch(Fuse, this.tableData, ['ip']);
     },
     sortChange(sortProps) {
       this.processData(sortProps, true);
     },
     sorting(sortProps, data) {
-      if (sortProps.column.label === 'IP Address' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.ip > b.ip) {
-            val = 1;
-          } else if (a.ip < b.ip) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'IP Address' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.ip < b.ip) {
-            val = 1;
-          } else if (a.ip > b.ip) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Total Application Running' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.count > b.apps.count) {
-            val = 1;
-          } else if (a.apps.count < b.apps.count) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Total Application Running' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.count < b.apps.count) {
-            val = 1;
-          } else if (a.apps.count > b.apps.count) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Flux Watch Tower Installed' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.fluxtower > b.apps.fluxtower) {
-            val = 1;
-          } else if (a.apps.fluxtower < b.apps.fluxtower) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Flux Watch Tower Installed' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.fluxtower < b.apps.fluxtower) {
-            val = 1;
-          } else if (a.apps.fluxtower > b.apps.fluxtower) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Flux Usage' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.fluxusage > b.apps.fluxusage) {
-            val = 1;
-          } else if (a.apps.fluxusage < b.apps.fluxusage) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Flux Usage' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.fluxusage < b.apps.fluxusage) {
-            val = 1;
-          } else if (a.apps.fluxusage > b.apps.fluxusage) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'CPU Locked' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.resources.appsCpusLocked > b.apps.resources.appsCpusLocked) {
-            val = 1;
-          } else if (a.apps.resources.appsCpusLocked < b.apps.resources.appsCpusLocked) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'CPU Locked' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.resources.appsCpusLocked < b.apps.resources.appsCpusLocked) {
-            val = 1;
-          } else if (a.apps.resources.appsCpusLocked > b.apps.resources.appsCpusLocked) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'RAM Locked' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.resources.appsRamLocked > b.apps.resources.appsRamLocked) {
-            val = 1;
-          } else if (a.apps.resources.appsRamLocked < b.apps.resources.appsRamLocked) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'RAM Locked' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.resources.appsRamLocked < b.apps.resources.appsRamLocked) {
-            val = 1;
-          } else if (a.apps.resources.appsRamLocked > b.apps.resources.appsRamLocked) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'HDD Locked' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.resources.appsHddLocked > b.apps.resources.appsHddLocked) {
-            val = 1;
-          } else if (a.apps.resources.appsHddLocked < b.apps.resources.appsHddLocked) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'HDD Locked' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.apps.resources.appsHddLocked < b.apps.resources.appsHddLocked) {
-            val = 1;
-          } else if (a.apps.resources.appsHddLocked > b.apps.resources.appsHddLocked) {
-            val = -1;
-          }
-          return val;
-        });
-      } else {
-        this.tableData = JSON.parse(this.originalData);
-      }
-      return data;
+      const ret = SortService.sortNodeApp(data, sortProps, this.originalData);
+      this.tableData = Object.keys(ret.tableDatas).length > 0 ? ret.tableDatas : this.tableData;
+      return ret.datas;
     },
     processDataForCsv(data) {
       const values = [];
@@ -536,33 +383,18 @@ export default {
       return values;
     },
     downloadCsvFile(data) {
-      const date = new Date();
-      const month = date.getMonth();
-      const day = date.getDate();
-      const year = date.getFullYear();
-      const options = {
-        filename: `Node_Application_${month}${day}${year}`,
-        fieldSeparator: ',',
-        quoteStrings: '"',
-        decimalSeparator: '.',
-        showLabels: true,
-        showTitle: true,
-        title: `Node Application - ${month}/${day}/${year}`,
-        useTextFile: false,
-        useBom: true,
-        headers: [
-          'IP Address',
-          'Total Application Running',
-          'Flux Tower Installed',
-          'Flux Usage',
-          'CPU Locked',
-          'RAM Locked',
-          'HDD Locked',
-          'Running Apps',
-        ],
-      };
-      const csvExporter = new ExportToCsv(options);
-      csvExporter.generateCsv(this.processDataForCsv(data));
+      const module = 'Node_Application';
+      const headers = [
+        'IP Address',
+        'Total Application Running',
+        'Flux Tower Installed',
+        'Flux Usage',
+        'CPU Locked',
+        'RAM Locked',
+        'HDD Locked',
+        'Running Apps',
+      ];
+      CsvService.Download(this.processDataForCsv(data), headers, module, ExportToCsv);
     },
     processFilters(key) {
       if (!this.filters.default.includes(key)) {
@@ -576,14 +408,7 @@ export default {
       return this.processData(false, false);
     },
     processState(keys) {
-      this.filters.states.map((item) => {
-        const values = item;
-        values.state = false;
-        if (keys.includes(values.name)) {
-          values.state = true;
-        }
-        return values;
-      });
+      this.filters = TransformationService.processState(keys, this.filters);
     },
   },
 };
