+++ ui/src/pages/Dashboard/NodeInfo/NodeLocation.vue
@@ -158,6 +158,10 @@ import axios from 'axios';
 import { MemoryStorage } from 'ttl-localstorage';
 import { ExportToCsv } from 'export-to-csv';
 import VueElementLoading from 'vue-element-loading';
+import CsvService from '../Service/CsvService';
+import SearchService from '../Service/SearchService';
+import TransformationService from '../Service/TransformationService';
+import SortService from '../Service/SortService';
 import {
   httpRequestFluxInfo, httpRequestDaemonInfo, httpRequestFluxHistoryStats,
 } from '../Request/HttpRequest';
@@ -281,25 +285,9 @@ export default {
     processData(sortProps, isProcessingState) {
       let result;
       if (this.searchQuery !== '') {
-        const temp = [];
-        result = this.fuseSearch.search(`=${this.searchQuery}`);
-        for (let i = 0; i < Object.keys(result).length; i += 1) {
-          temp.push(result[i].item);
-        }
-        result = temp;
+        result = SearchService.search(this.fuseSearch, this.searchQuery);
       } else if (this.filters.default.length) {
-        const arr = [];
-        const data = [];
-        this.filters.default.forEach((item) => {
-          const objs = this.filter.get(item);
-          objs.forEach((obj) => {
-            if (!arr.includes(`${obj.geolocation.ip}${obj.geolocation.country}${obj.node.status.tier}`)) {
-              arr.push(`${obj.geolocation.ip}${obj.geolocation.country}${obj.node.status.tier}`);
-              data.push(obj);
-            }
-          });
-        });
-        result = data;
+        result = TransformationService.processFilters(this.filters, this.filter, 'nodelocation');
       } else {
         result = this.tableData;
       }
@@ -385,196 +373,15 @@ export default {
     },
     setSearch() {
       this.originalData = JSON.stringify(this.tableData);
-      this.fuseSearch = new Fuse(this.tableData, { useExtendedSearch: true, keys: ['geolocation.ip'] });
+      this.fuseSearch = SearchService.generateSearch(Fuse, this.tableData, ['geolocation.ip']);
     },
     sortChange(sortProps) {
       this.processData(sortProps, true);
     },
     sorting(sortProps, data) {
-      if (sortProps.column.label === 'IP Address' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.ip > b.geolocation.ip) {
-            val = 1;
-          } else if (a.geolocation.ip < b.geolocation.ip) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'IP Address' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.ip < b.geolocation.ip) {
-            val = 1;
-          } else if (a.geolocation.ip > b.geolocation.ip) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Country' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.country > b.geolocation.country) {
-            val = 1;
-          } else if (a.geolocation.country < b.geolocation.country) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Country' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.country < b.geolocation.country) {
-            val = 1;
-          } else if (a.geolocation.country > b.geolocation.country) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Country Code' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.countryCode > b.geolocation.countryCode) {
-            val = 1;
-          } else if (a.geolocation.countryCode < b.geolocation.countryCode) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Country Code' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.countryCode < b.geolocation.countryCode) {
-            val = 1;
-          } else if (a.geolocation.countryCode > b.geolocation.countryCode) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Region' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.region > b.geolocation.region) {
-            val = 1;
-          } else if (a.geolocation.region < b.geolocation.region) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Region' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.region < b.geolocation.region) {
-            val = 1;
-          } else if (a.geolocation.region > b.geolocation.region) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Region Name' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.regionName > b.geolocation.regionName) {
-            val = 1;
-          } else if (a.geolocation.regionName < b.geolocation.regionName) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Region Name' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.regionName < b.geolocation.regionName) {
-            val = 1;
-          } else if (a.geolocation.regionName > b.geolocation.regionName) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Latitude' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.lat > b.geolocation.lat) {
-            val = 1;
-          } else if (a.geolocation.lat < b.geolocation.lat) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Latitude' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.lat < b.geolocation.lat) {
-            val = 1;
-          } else if (a.geolocation.lat > b.geolocation.lat) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Longitude' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.lon > b.geolocation.lon) {
-            val = 1;
-          } else if (a.geolocation.lon < b.geolocation.lon) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Longitude' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.lon < b.geolocation.lon) {
-            val = 1;
-          } else if (a.geolocation.lon > b.geolocation.lon) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Organization' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.org > b.geolocation.org) {
-            val = 1;
-          } else if (a.geolocation.org < b.geolocation.org) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Organization' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.geolocation.org < b.geolocation.org) {
-            val = 1;
-          } else if (a.geolocation.org > b.geolocation.org) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Tier' && sortProps.column.order === 'ascending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.node.status.tier > b.node.status.tier) {
-            val = 1;
-          } else if (a.node.status.tier < b.node.status.tier) {
-            val = -1;
-          }
-          return val;
-        });
-      } else if (sortProps.column.label === 'Tier' && sortProps.column.order === 'descending') {
-        data.sort((a, b) => {
-          let val = 0;
-          if (a.node.status.tier < b.node.status.tier) {
-            val = 1;
-          } else if (a.node.status.tier > b.node.status.tier) {
-            val = -1;
-          }
-          return val;
-        });
-      } else {
-        this.tableData = JSON.parse(this.originalData);
-      }
-      return data;
+      const ret = SortService.sortNodeLocation(data, sortProps, this.originalData);
+      this.tableData = Object.keys(ret.tableDatas).length > 0 ? ret.tableDatas : this.tableData;
+      return ret.datas;
     },
     processDataForCsv(data) {
       const values = [];
@@ -593,33 +400,18 @@ export default {
       return values;
     },
     downloadCsvFile(data) {
-      const date = new Date();
-      const month = date.getMonth();
-      const day = date.getDate();
-      const year = date.getFullYear();
-      const options = {
-        filename: `Node_Geolocation_${month}${day}${year}`,
-        fieldSeparator: ',',
-        quoteStrings: '"',
-        decimalSeparator: '.',
-        showLabels: true,
-        showTitle: true,
-        title: `Node Geolocation - ${month}/${day}/${year}`,
-        useTextFile: false,
-        useBom: true,
-        headers: [
-          'IP Address',
-          'Country',
-          'Country Code',
-          'Region',
-          'Region Name',
-          'Latitude',
-          'Longitude',
-          'Tier',
-        ],
-      };
-      const csvExporter = new ExportToCsv(options);
-      csvExporter.generateCsv(this.processDataForCsv(data));
+      const module = 'Node_Geolocation';
+      const headers = [
+        'IP Address',
+        'Country',
+        'Country Code',
+        'Region',
+        'Region Name',
+        'Latitude',
+        'Longitude',
+        'Tier',
+      ];
+      CsvService.Download(this.processDataForCsv(data), headers, module, ExportToCsv);
     },
     processFilters(key) {
       if (!this.filters.default.includes(key)) {
@@ -633,14 +425,7 @@ export default {
       return this.processData(false, false);
     },
     processState(keys) {
-      this.filters.states.map((item) => {
-        const values = item;
-        values.state = false;
-        if (keys.includes(values.name)) {
-          values.state = true;
-        }
-        return values;
-      });
+      this.filters = TransformationService.processState(keys, this.filters);
     },
   },
 };
